+++ app/src/main/java/com/belcobtm/domain/referral/item/ReferralDataItem.kt
@@ -4,5 +4,5 @@ data class ReferralDataItem(
     val link: String,
     val message: String,
     val invited: Int,
-    val earned: Int
+    val earned: Double
 )
