+++ app/src/main/java/com/belcobtm/domain/transaction/item/TransactionPlanItem.kt
@@ -1,13 +1,14 @@
 package com.belcobtm.domain.transaction.item
 
 import com.belcobtm.data.rest.transaction.response.hash.TronBlockHeaderResponse
+import java.math.BigInteger
 
 data class TransactionPlanItem(
     val coinCode: String,
     val byteFee: Long,
-    val nonce: Int,
-    val gasPrice: Int,
-    val gasLimit: Int,
+    val nonce: BigInteger,
+    val gasPrice: BigInteger,
+    val gasLimit: BigInteger,
     val txFee: Double,
     val nativeTxFee: Double,
     val blockHeader: TronBlockHeaderResponse?,
